+++ DeviceSelector/translations/DeviceSelector_ko.ts
@@ -1,0 +1,350 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS>
+<TS version="2.1" language="ko_KR">
+<context>
+    <name>DeviceSelector</name>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="43"/>
+        <source>Playback devices</source>
+        <translation>재생 장치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="48"/>
+        <source>Capture devices</source>
+        <translation>녹음 장치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="55"/>
+        <location filename="../DeviceSelector.cpp" line="187"/>
+        <location filename="../main.cpp" line="67"/>
+        <source>Error while accessing the registry</source>
+        <translation>레지스트리에 접근하는 중 오류가 발생했습니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="99"/>
+        <location filename="../DeviceSelector.cpp" line="222"/>
+        <source>Info</source>
+        <translation>알림</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="99"/>
+        <source>A registry value that is required for the operation of Equalizer APO was not set correctly. This might have been caused by a driver installation or uninstallation. The value has been corrected. A reboot may be required so that the changes can take effect.</source>
+        <translation>Equalizer APO 동작에 필요한 레지스트리 값이 올바르게 설정되어 있지 않았습니다. 드라이버 설치나 제거 때문일 수 있습니다. 값을 바로잡았으며, 변경 사항을 적용하려면 재부팅이 필요할 수 있습니다.</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="201"/>
+        <source>Upgrades available</source>
+        <translation>업그레이드 가능</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="201"/>
+        <source>The APO installation of some devices should be upgraded. Do you really want to cancel?</source>
+        <translation>일부 장치의 APO 설치를 업그레이드해야 합니다. 정말 취소하시겠습니까?</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="222"/>
+        <source>This dialog can be reopened anytime by launching Device Selector from the start menu.</source>
+        <translation>이 창은 시작 메뉴에서 Device Selector를 실행해 언제든지 다시 열 수 있습니다.</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="228"/>
+        <source>Reboot</source>
+        <translation>재부팅</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="228"/>
+        <source>To apply the changes, Windows should be rebooted. Reboot now?</source>
+        <translation>변경 사항을 적용하려면 Windows를 재부팅해야 합니다. 지금 재부팅하시겠습니까?</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="343"/>
+        <source>Cancel</source>
+        <translation>취소</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="350"/>
+        <source>Close</source>
+        <translation>닫기</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="481"/>
+        <source>APO will be installed</source>
+        <translation>APO를 설치합니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="483"/>
+        <source>APO will be uninstalled</source>
+        <translation>APO를 제거합니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="485"/>
+        <source>APO will be upgraded</source>
+        <translation>APO를 업그레이드합니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="487"/>
+        <source>APO installation will be changed</source>
+        <translation>APO 설치를 변경합니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="489"/>
+        <source>Audio enhancements will be enabled</source>
+        <translation>오디오 향상 기능을 켭니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="491"/>
+        <source>APO is already installed</source>
+        <translation>APO가 이미 설치되어 있습니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="493"/>
+        <source>APO can be installed (experimental)</source>
+        <translation>APO를 설치할 수 있습니다 (실험적)</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="495"/>
+        <source>APO can be installed</source>
+        <translation>APO를 설치할 수 있습니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="499"/>
+        <source>Voicemeeter was uninstalled</source>
+        <translation>Voicemeeter가 제거되었습니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="501"/>
+        <source>Default device</source>
+        <translation>기본 장치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="504"/>
+        <source>Disabled</source>
+        <translation>사용 안 함</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.cpp" line="506"/>
+        <source>Unplugged</source>
+        <translation>연결 끊김</translation>
+    </message>
+</context>
+<context>
+    <name>DeviceSelectorClass</name>
+    <message>
+        <location filename="../DeviceSelector.ui" line="27"/>
+        <source>Troubleshooting options (only use in case of problems)</source>
+        <translation>문제 해결 옵션 (문제가 있을 때만 사용)</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="58"/>
+        <source>Please select exactly one activated device by clicking on its name</source>
+        <translation>이름을 클릭해 활성화된 장치를 하나만 선택하십시오</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="72"/>
+        <location filename="../DeviceSelector.ui" line="102"/>
+        <source>Install APO</source>
+        <translation>APO 설치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="79"/>
+        <location filename="../DeviceSelector.ui" line="160"/>
+        <source>Use original APO</source>
+        <translation>원본 APO 사용</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="92"/>
+        <source>Pre-mix:</source>
+        <translation>믹스 전:</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="131"/>
+        <source>Post-mix:</source>
+        <translation>믹스 후:</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="142"/>
+        <source>Install as LFX/GFX</source>
+        <translation>LFX/GFX로 설치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="147"/>
+        <source>Install as SFX/MFX</source>
+        <translation>SFX/MFX로 설치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="152"/>
+        <source>Install as SFX/EFX</source>
+        <translation>SFX/EFX로 설치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="167"/>
+        <source>Automatically adjust install mode if problems with current mode are detected</source>
+        <translation>현재 모드에 문제가 감지되면 설치 모드를 자동으로 조정</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="170"/>
+        <source>Auto:</source>
+        <translation>자동:</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="177"/>
+        <source>Allows to change silent input buffers to audible output buffers, so that delayed audio can be played.
+
+CAUTION: Some sound card drivers can produce unwanted sounds when this is enabled.</source>
+        <translation>무음 입력 버퍼를 소리가 나는 출력 버퍼로 바꿀 수 있게 하여, 지연된 오디오를 재생할 수 있습니다.
+
+주의: 일부 사운드 카드 드라이버에서는 이 옵션을 켜면 원치 않는 소리가 날 수 있습니다.</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="182"/>
+        <source>Allow silent buffer modification</source>
+        <translation>무음 버퍼 수정 허용</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="209"/>
+        <source>Connector</source>
+        <translation>커넥터</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="214"/>
+        <source>Device</source>
+        <translation>장치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="219"/>
+        <source>Status</source>
+        <translation>상태</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="234"/>
+        <source>Please select the devices for which Equalizer APO is to be installed:</source>
+        <translation>Equalizer APO를 설치할 장치를 선택하십시오:</translation>
+    </message>
+    <message>
+        <location filename="../DeviceSelector.ui" line="241"/>
+        <source>Copy Device command to clipboard</source>
+        <translation>Device 명령을 클립보드에 복사</translation>
+    </message>
+</context>
+<context>
+    <name>DeviceTestDialog</name>
+    <message>
+        <location filename="../DeviceTestDialog.cpp" line="42"/>
+        <source>Playback devices</source>
+        <translation>재생 장치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestDialog.cpp" line="51"/>
+        <source>Capture devices</source>
+        <translation>녹음 장치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestDialog.cpp" line="59"/>
+        <source>Error while accessing the registry</source>
+        <translation>레지스트리에 접근하는 중 오류가 발생했습니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestDialog.cpp" line="124"/>
+        <source>Cannot test APO installation as device is disabled</source>
+        <translation>장치가 사용 안 함 상태라 APO 설치를 테스트할 수 없습니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestDialog.cpp" line="129"/>
+        <source>Cannot test APO installation as device is unplugged</source>
+        <translation>장치가 연결 끊김 상태라 APO 설치를 테스트할 수 없습니다</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestDialog.cpp" line="167"/>
+        <location filename="../DeviceTestDialog.cpp" line="213"/>
+        <source>Error</source>
+        <translation>오류</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestDialog.cpp" line="187"/>
+        <source>Equalizer APO works but original APO could not be initialized. Maybe unset &quot;Use original APO&quot; in troubleshooting options.</source>
+        <translation>Equalizer APO는 동작하지만 원본 APO를 초기화하지 못했습니다. 문제 해결 옵션에서 "원본 APO 사용"을 해제해 보십시오.</translation>
+    </message>
+</context>
+<context>
+    <name>DeviceTestDialogClass</name>
+    <message>
+        <location filename="../DeviceTestDialog.ui" line="14"/>
+        <source>Testing APO installation on devices...</source>
+        <translation>장치에서 APO 설치를 테스트하는 중...</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestDialog.ui" line="37"/>
+        <source>Connector</source>
+        <translation>커넥터</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestDialog.ui" line="42"/>
+        <source>Device</source>
+        <translation>장치</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestDialog.ui" line="47"/>
+        <source>Pre-mix</source>
+        <translation>믹스 전</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestDialog.ui" line="52"/>
+        <source>Post-mix</source>
+        <translation>믹스 후</translation>
+    </message>
+</context>
+<context>
+    <name>DeviceTestThread</name>
+    <message>
+        <location filename="../DeviceTestThread.cpp" line="69"/>
+        <location filename="../DeviceTestThread.cpp" line="230"/>
+        <source>Restarting audio service...</source>
+        <translation>오디오 서비스를 다시 시작하는 중...</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestThread.cpp" line="74"/>
+        <location filename="../DeviceTestThread.cpp" line="235"/>
+        <source>Restart failed.</source>
+        <translation>다시 시작하지 못했습니다.</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestThread.cpp" line="88"/>
+        <source>Checking APO installation...</source>
+        <translation>APO 설치를 확인하는 중...</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestThread.cpp" line="136"/>
+        <source>Received unknown device GUID %1.</source>
+        <translation>알 수 없는 장치 GUID %1을(를) 받았습니다.</translation>
+    </message>
+    <message numerus="yes">
+        <location filename="../DeviceTestThread.cpp" line="172"/>
+        <source>Check failed for %n device(s).</source>
+        <translation>
+            <numerusform>장치 %n개에서 확인에 실패했습니다.</numerusform>
+        </translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestThread.cpp" line="217"/>
+        <source>Setting install mode for %1 %2 to %3.</source>
+        <translation>%1 %2의 설치 모드를 %3(으)로 설정합니다.</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestThread.cpp" line="226"/>
+        <source>Trying other configurations...</source>
+        <translation>다른 구성을 시도하는 중...</translation>
+    </message>
+    <message>
+        <location filename="../DeviceTestThread.cpp" line="249"/>
+        <source>Checks done. No problems were detected.</source>
+        <translation>확인을 마쳤습니다. 문제가 발견되지 않았습니다.</translation>
+    </message>
+    <message numerus="yes">
+        <location filename="../DeviceTestThread.cpp" line="251"/>
+        <source>Checks done. Problems were detected for %n device(s).</source>
+        <translation>
+            <numerusform>확인을 마쳤습니다. 장치 %n개에서 문제가 발견되었습니다.</numerusform>
+        </translation>
+    </message>
+</context>
+</TS>